handle error a node

+++ src/main/java/com/sapo/mock/techshop/service/impl/ActivityManager.java
@@ -1,6 +1,5 @@
 package com.sapo.mock.techshop.service.impl;
 
-import com.sapo.mock.techshop.schedules.TestJob;
 import com.sapo.mock.techshop.service.TestService;
 import lombok.RequiredArgsConstructor;
 import org.apache.curator.framework.CuratorFramework;
@@ -18,40 +17,53 @@ public class ActivityManager {
     public static final String ACTIVITY_PATH = "/schedule/";
     public static final String ERROR_PATH = "/error/";
 
-    public void performActivity(String activityName) throws Exception {
+    public void performActivity(String activityName, boolean retry) throws Exception {
         String activityPath = ACTIVITY_PATH + activityName;
         String errorPath = ERROR_PATH + activityName;
 
-        this.createPathIfNeeded(activityPath);
+        try {
+            if (retry) {
+                client.delete().forPath(activityPath);
+            }
+        } catch (Exception e) {
+            System.out.println("Retry on another node");
+        }
 
         try {
-            switch (activityName) {
-                case "scheduleTask":
-                    testService.test("1");
+            if (this.createPathIfNeeded(activityPath)) {
+                switch (activityName) {
+                    case "scheduleTask":
+                        testService.test("1");
+                }
+                client.delete().forPath(activityPath);
+                this.resetCounter(errorPath);
+            } else {
+                System.out.println("Job runing on another node");
             }
-            client.delete().forPath(activityPath);
-            this.resetCounter(errorPath);
         } catch (Exception e) {
             // Tăng counter nếu xảy ra lỗi
             this.incrementErrorCounter(errorPath);
         }
     }
 
-    public void createPathIfNeeded(String path) throws Exception {
-        if (client.checkExists().forPath(path) == null) {
+    public boolean createPathIfNeeded(String path) {
+        try {
             client.create().creatingParentsIfNeeded().forPath(path);
+            return true;
+        } catch (Exception e) {
+            return false;
         }
     }
 
     public void incrementErrorCounter(String errorPath) throws Exception {
         DistributedAtomicInteger counter = new DistributedAtomicInteger(
-                client, errorPath, new RetryNTimes(10, 1000));
+                client, errorPath, new RetryNTimes(0, 1000));
         counter.increment();
     }
 
     public void resetCounter(String errorPath) throws Exception {
         DistributedAtomicInteger counter = new DistributedAtomicInteger(
-                client, errorPath, new RetryNTimes(10, 1000));
+                client, errorPath, new RetryNTimes(0, 1000));
         counter.forceSet(0);
     }
 }
